# docs/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TransEV EV CMS API
  version: 0.11.0
  description: |
    Canonical machine-readable contract for the endpoints currently implemented
    by `ev-cms-backend-new`.

    Authentication uses a signed-then-encrypted bearer access token. Platform
    sessions and CPO sessions are separate. CPO integration operations also
    require `X-CPO-App-ID`; that header is checked only after the bearer token
    establishes the CPO and never selects a tenant.

    Public customer signup requires `X-CPO-App-ID` to resolve an active CPO, but
    the header remains public routing metadata rather than authentication.
    Administrative authentication, platform control-plane, health, and
    documentation routes do not require it. Request bodies reject unknown
    properties and are limited to 32 KiB by the server.

    Every request that reaches Gin returns a server-generated UUID in the
    `X-Request-ID` response header for correlation with the safe structured
    completion log. The logger records matched route templates and stable
    handled error codes, never bodies, raw paths, queries, credentials, tokens,
    OTPs, email addresses, user agents, app IDs, API messages, or panic values.
    `LOG_LEVEL=DEBUG` adds request-start and handled-error component/type
    diagnostics under the same request ID without relaxing those exclusions.
  contact:
    name: TransEV Engineering
servers:
  - url: /
    description: Same origin as the API documentation page
...
paths:
...
  /api/v1/platform/cpos:
    post:
      tags: [Platform CPOs]
      operationId: createCPO
      summary: Create a pending CPO and its first administrator
      description: |
        Atomically creates a pending CPO with no commercial prerequisite, a
        unique dummy app ID, an ADMIN membership, audit data, and an encrypted
        onboarding mail job. A new email receives a generated temporary
        password; an existing active identity keeps its password. The
        temporary password never appears in the response. The transaction
        fails rather than queueing a new-identity welcome without that
        credential; a successful response proves the job committed, not that
        SMTP has sent it.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCPORequest"
      responses:
        "201":
          description: CPO and first administrator created or attached.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCPOResponse"
        "400":
          description: |
            `invalid_request`, `invalid_slug`, `invalid_business_name`,
            `invalid_company_type`, `invalid_gstin`, `invalid_address`,
            `invalid_city`, `invalid_state`, `invalid_pincode`,
            `invalid_admin_email`, or `invalid_admin_full_name`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: |
            `cpo_slug_conflict`, `cpo_gstin_conflict`,
            `cpo_app_id_conflict`, `admin_identity_conflict`,
            `cpo_admin_membership_conflict`, `cpo_primary_admin_conflict`,
            `admin_identity_inactive`, or the fallback `cpo_conflict`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          $ref: "#/components/responses/MailUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags: [Platform CPOs]
      operationId: listCPOs
      summary: Search and cursor-page CPOs
      description: |
        Returns a stable newest-first page. Search matches business name, slug,
        GSTIN, app ID, and the designated primary administrator's name/email.
        When `has_more` is true, send both returned cursor fields unchanged.
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
            maxLength: 200
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, ACTIVE, SUSPENDED]
        - name: app_id_mode
          in: query
          schema:
            $ref: "#/components/schemas/CPOAppIDMode"
        - name: before
          in: query
          description: Exclusive RFC3339 creation-time cursor; requires before_id.
          schema:
            type: string
            format: date-time
        - name: before_id
          in: query
          description: Exclusive UUID tie-breaker; requires before.
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: CPO collection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CPOListResponse"
        "400":
          description: |
            `invalid_q`, `invalid_status`, `invalid_app_id_mode`,
            `invalid_before`, `invalid_before_id`, `invalid_cursor`, or
            `invalid_limit`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/platform/cpos/slug-availability:
    get:
      tags: [Platform CPOs]
      operationId: checkCPOSlugAvailability
      summary: Check whether a normalized CPO slug is currently available
      description: |
        Trims and lowercases the candidate, validates the same format used by
        CPO creation, and reports whether the normalized slug is currently
        unallocated. This read is advisory: concurrent creation can consume an
        available slug, so POST /api/v1/platform/cpos remains authoritative and
        can still return cpo_slug_conflict.
      security:
        - BearerAuth: []
      parameters:
        - name: slug
          in: query
          required: true
          description: Candidate CPO slug. The server trims and lowercases it.
          schema:
            type: string
            maxLength: 80
            pattern: "^[A-Za-z0-9]+(?:-[A-Za-z0-9]+)*$"
      responses:
        "200":
          description: Current availability of the normalized candidate.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CPOSlugAvailability"
        "400":
          description: "`invalid_slug`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/platform/cpos/{cpo_id}:
    get:
      tags: [Platform CPOs]
      operationId: getCPO
      summary: Get one CPO
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CPOID"
      responses:
        "200":
          description: CPO record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CPO"
        "400":
          $ref: "#/components/responses/InvalidCPOID"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/CPONotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/platform/cpos/{cpo_id}/profile:
    put:
      tags: [Platform CPOs]
      operationId: updateCPOProfile
      summary: Replace the mutable CPO business profile
      description: |
        Replaces business-profile fields in one transaction and emits audit and
        realtime evidence. The stable slug, lifecycle, app ID, and tenant data
        are not changed.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CPOID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCPOProfileRequest"
      responses:
        "200":
          description: Updated CPO.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CPO"
        "400":
          description: |
            `invalid_request`, `invalid_cpo_id`, `invalid_business_name`,
            `invalid_company_type`, `invalid_gstin`, `invalid_address`,
            `invalid_city`, `invalid_state`, or `invalid_pincode`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/CPONotFound"
        "409":
          description: "`cpo_gstin_conflict` — GSTIN is already assigned."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/platform/cpos/{cpo_id}/activate:
    post:
      tags: [Platform CPOs]
      operationId: activateCPO
      summary: Set a CPO active
      description: |
        Idempotently grants platform access by setting the CPO lifecycle status
        to ACTIVE. A live app ID is not required. CPO administrative login
        becomes eligible.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CPOID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReasonRequest"
      responses:
        "200":
          description: Current active CPO record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CPO"
        "400":
          description: "`invalid_request`, `invalid_cpo_id`, or `invalid_reason`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/CPONotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/platform/cpos/{cpo_id}/suspend:
    post:
      tags: [Platform CPOs]
      operationId: suspendCPO
      summary: Suspend a CPO and revoke its sessions
      description: |
        Idempotently sets status to SUSPENDED and revokes every active CPO-staff
        and customer session plus unused refresh tokens for the tenant. New
        tenant login is blocked. Platform sessions are unaffected.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CPOID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReasonRequest"
      responses:
        "200":
          description: Current suspended CPO record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CPO"
        "400":
          description: "`invalid_request`, `invalid_cpo_id`, or `invalid_reason`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/CPONotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/cpo/organization:
    get:
      tags: [CPO Operations]
      operationId: getCPOOrganization
      summary: Read the authenticated CPO's organization details
      description: |
        Returns a tenant-safe, read-only projection of the CPO registration
        and operational identity managed by the platform. The CPO ID is
        derived from the verified ADMIN session; callers cannot select another
        tenant. Internal platform actor IDs and the privileged lifecycle reason
        are omitted. Organization changes remain Superadmin-only.
      security:
        - BearerAuth: []
          CPOAppID: []
      responses:
        "200":
          description: Current CPO organization details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CPOOrganization"
        "400":
          $ref: "#/components/responses/MissingCPOAppID"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/CPOBusinessForbidden"
        "404":
          $ref: "#/components/responses/CPONotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWE
      description: Short-lived signed-then-encrypted administrative access token.
    CPOAppID:
      type: apiKey
      in: header
      name: X-CPO-App-ID
      description: |
        Current app ID for the CPO already established by BearerAuth. This is
        routing metadata, not an authentication secret or tenant selector.
...
  schemas:
...
    CPOAppIDMode:
      type: string
      enum: [DUMMY, LIVE]
...